Fix correctness bugs and refactor for encapsulation
Bugs:
- Lap display >= 1h now uses HH:MM:SS (was silently truncated to MM:SS.cs)
- Timer completion uses remaining().is_zero() (was firing up to ~1s early)
- RAII TerminalGuard restores terminal on any error return path
- Remove dead end_time computation each frame
- "Ends at" honors 24h preference; shows "Complete" when finished
- adjust_minutes/seconds keep a running timer running (recompute end_time only)

Refactors:
- Extract App::handle_key; tests call the real dispatch
- Encapsulate CountdownTimer/Stopwatch/SystemNotifier fields behind accessors
- Add explicit is_finished flag to CountdownTimer (don't overload is_paused)
- Consolidate config defaults into single default_raw_config() source
- Add fallible parse_color returning Result<Color, ColorParseError>
- Switch signal.rs to signal_hook::flag::register (no unsafe)

Docs: update AGENTS.md and README.md with new behavior and test counts
(141 defined / 138 cross-platform; clippy clean)

+++ src/ui.rs
@@ -19,8 +19,8 @@ pub fn create_main_layout(area: Rect) -> Vec<Rect> {
 }
 
 pub fn render_large_text(f: &mut Frame, area: Rect, text: &str, color: Color, font: &LargeFont) {
-    let base_w = font.glyph_width() as usize;
-    let base_h = font.glyph_height() as usize;
+    let base_w = font.glyph_width();
+    let base_h = font.glyph_height();
     let text_chars: Vec<char> = text.chars().collect();
     let num_chars = text_chars.len();
